+++ R/netcomb.R
@@ -81,8 +81,8 @@
 #' \item{TE}{Estimate of treatment effect, i.e. difference between
 #'   first and second treatment.}
 #' \item{seTE}{Standard error of treatment estimate.}
-#' \item{seTE.adj}{Standard error of treatment estimate, adjusted for
-#'   multi-arm studies.}
+#' \item{seTE.adj.fixed, seTE.adj.random}{Standard error of treatment
+#'   estimate, adjusted for multi-arm studies.}
 #' \item{design}{Design of study providing pairwise comparison.}
 #' \item{event1}{Number of events in first treatment group.}
 #' \item{event2}{Number of events in second treatment group.}
@@ -543,6 +543,8 @@ netcomb <- function(x,
               TE = x$TE,
               seTE = x$seTE,
               seTE.adj = x$seTE.adj,
+              seTE.adj.fixed = x$seTE.adj.fixed,
+              seTE.adj.random = x$seTE.adj.random,
               ##
               design = x$design,
               ##