+++ R/netrank.R
@@ -72,7 +72,7 @@
 #' \item{version}{Version of R package netmeta used to create object.}
 #' 
 #' @author Gerta Rücker \email{ruecker@@imbi.uni-freiburg.de}, Guido
-#'   Schwarzer \email{sc@@imbi.uni-freiburg.de}, Thodoris
+#'   Schwarzer \email{sc@@imbi.uni-freiburg.de}, Theodoros
 #'   Papakonstantinou \email{dev@tpapak.com}
 #' 
 #' @seealso \code{\link{netmeta}}, \code{\link{rankogram}}